rename mode to sequence
