+++ src/lib/agent/agent-interface.ts
@@ -161,7 +161,7 @@ export type AgentConfig = {
    * flag routes the run here; threaded into wizard-tools so the orchestrator
    * tools register.
    */
-  orchestrator?: import('@lib/agent/runner/mode/orchestrator/queue-tools').OrchestratorToolsContext;
+  orchestrator?: import('@lib/agent/runner/sequence/orchestrator/queue-tools').OrchestratorToolsContext;
 };
 
 /**
