+++ src/lib/agent/agent-prompt-loader.ts
@@ -15,8 +15,11 @@
  * network latency. The registry's type list also drives `enqueue_task`
  * validation.
  */
-import type { QueueStore, QueuedTask } from './runner/mode/orchestrator/queue';
-import type { ResolvedTask } from './runner/mode/orchestrator/executor';
+import type {
+  QueueStore,
+  QueuedTask,
+} from './runner/sequence/orchestrator/queue';
+import type { ResolvedTask } from './runner/sequence/orchestrator/executor';
 import { DEFAULT_AGENT_MODEL } from '@lib/constants';
 
 /**
