+++ src/lib/wizard-tools.ts
@@ -29,7 +29,7 @@ import { createSecretVault, type SecretVault } from './secret-vault';
 import {
   buildOrchestratorTools,
   type OrchestratorToolsContext,
-} from './agent/runner/mode/orchestrator/queue-tools';
+} from './agent/runner/sequence/orchestrator/queue-tools';
 
 // ---------------------------------------------------------------------------
 // SDK dynamic import (ESM module loaded once, cached)
